Create a testutils/test_reporter.py file
